Fix the binary compatibility
